requirements.txt
greenlet==1.1.0
importlib-metadata==4.0.1
python-dateutil==2.8.1
six==1.16.0
SQLAlchemy==1.4.15
tornado==6.1
typing-extensions==3.10.0.0
zipp==3.4.1
requests~=2.25.1